lcb: Implemented constant sequences (#91)
lcb: Implemented constant sequences

+++ vadl/main/vadl/gcb/passes/ValueRange.java
@@ -19,5 +19,5 @@ package vadl.gcb.passes;
 /**
 * Indicates what value range an immediate can have.
 */
-public record ValueRange(int lowest, int highest) {
+public record ValueRange(long lowest, long highest) {
 }
